refactor(checkin): delegate user validation to command handler

+++ lotusplus/src/main/java/com/example/lotusplus/checkin/command/dto/CheckInResponse.java
@@ -20,6 +20,6 @@ public class CheckInResponse {
     /**
      * Tổng điểm Lotus+ sau khi check-in
      */
-    private Integer totalPoint;
+    private Long totalPoint;
 
 }
